refactor, impl IsIn for BigInteger
