+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/numeric/BigIntegerStorage.java
@@ -14,6 +14,7 @@ import org.enso.table.data.column.operation.map.numeric.comparisons.GreaterCompa
 import org.enso.table.data.column.operation.map.numeric.comparisons.GreaterOrEqualComparison;
 import org.enso.table.data.column.operation.map.numeric.comparisons.LessComparison;
 import org.enso.table.data.column.operation.map.numeric.comparisons.LessOrEqualComparison;
+import org.enso.table.data.column.operation.map.numeric.isin.BigIntegerIsInOp;
 import org.enso.table.data.column.storage.ObjectStorage;
 import org.enso.table.data.column.storage.SpecializedStorage;
 import org.enso.table.data.column.storage.type.BigIntegerType;
@@ -43,7 +44,8 @@ public class BigIntegerStorage extends SpecializedStorage<BigInteger> {
         .add(new LessOrEqualComparison<>())
         .add(new EqualsComparison<>())
         .add(new GreaterOrEqualComparison<>())
-        .add(new GreaterComparison<>());
+        .add(new GreaterComparison<>())
+        .add(new BigIntegerIsInOp<>());
   }
 
   public static BigIntegerStorage makeEmpty(int size) {
